+++ CHANGELOG.md
@@ -78,6 +78,8 @@ BUG FIXES:
   before a "/" [GH-716]
 * core: "{{timestamp}}" values will always be the same for the entire
   duration of a build. [GH-744]
+* builder/amazon: Handle cases where security group isn't instantly
+  available. [GH-494]
 * builder/virtualbox: don't download guest additions if disabled. [GH-731]
 * post-processor/vsphere: Uploads VM properly. [GH-694]
 * post-processor/vsphere: Process user variables.